moved functions to proper files
