+++ src/app/login/login.page.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { AngularFireAuth } from '@angular/fire/auth';
+import { ToastController } from '@ionic/angular';
 
 @Component({
   selector: 'app-login',
@@ -7,7 +9,26 @@ import { Component, OnInit } from '@angular/core';
 })
 export class LoginPage implements OnInit {
 
-  constructor() { }
+  constructor(
+    public auth: AngularFireAuth,
+    public toastCtl: ToastController,
+    ) {}
+
+  login(email: string, password: string): void {
+    this.auth.signInWithEmailAndPassword(email, password).then(() => {
+      // the user is signed in
+    })
+    .catch((err) => {
+      this.createErrorToast(err.message);
+    });
+  }
+
+  async createErrorToast(msg: string) {
+    const toast = await this.toastCtl.create({
+      message: msg,
+      duration: 3000,
+    });
+  }
 
   ngOnInit() {
   }
